final?
so finnicky to the point where any wrong outcomes seem to be related to things out of my control like the way scripts detect distance, or the way unity works. E.g i start at a checkpoint in engine and it works fine, then later when i respawn at that checkpoint in the actual level it's different. The checkpoint marker and the respawn point are exactly the same spot.